v2/requirements.txt
streamlit>=1.28.0
google-genai>=0.3.0
python-dotenv>=1.0.0
mcp==1.26.0